requirements.txt
Django>=4.1.5
django-cors-headers>=3.13.0
djangorestframework>=3.14.0
requests>=2.28.0
pillow>=9.4.0
psycopg2-binary